LayoutTests/fast/dom/Element/scrollTop-expected.txt

Tests that for standard mode document.documentElement.scrollTop returns the scroll top value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.pageYOffset is 500
PASS document.body.scrollTop is 500
PASS document.documentElement.scrollTop is 0
PASS successfullyParsed is true

TEST COMPLETE
